Update predict.py
Add multiclass_dice_tf function. This calculates the dice loss score between the predicted multiclass mask and the real multiclass mask.

The prediction is returned in soft, probabilistic values, so it ensures they are converted to hard (ohe) values like the real data before testing dice.

Added an output for the dice score for testing. Achieving 0.86 on a trained model for 8 epochs.